Bulk edit on various things
In Class Quiz:
Ability to edit and delete the in class quiz is now added

In Class Quiz is now appearing in Performance page for teacher, student and parent

Added a chart for In Class Quiz and a chart for the Online MCQ Quizzes to show the percentage score of the student in the quizzes and show it as graph.
This feature is implemented for teacher view, Student view and Parent view as well.